+++ src/ctr_fatfs_interface.c
@@ -36,7 +36,7 @@ void ctr_fatfs_interface_destroy(ctr_fatfs_interface *io)
 	*io = (ctr_fatfs_interface){0};
 }
 
-int ctr_fatfs_interface_read(void *io, void *buffer, size_t buffer_size, size_t position, size_t count)
+int ctr_fatfs_interface_read(void *io, void *buffer, size_t buffer_size, uint64_t position, size_t count)
 {
 	FRESULT result = FR_OK;
 	if (count)
@@ -54,7 +54,7 @@ int ctr_fatfs_interface_read(void *io, void *buffer, size_t buffer_size, size_t 
 	return result != FR_OK;
 }
 
-int ctr_fatfs_interface_write(void *io, const void *buffer, size_t buffer_size, size_t position)
+int ctr_fatfs_interface_write(void *io, const void *buffer, size_t buffer_size, uint64_t position)
 {
 	FRESULT result = FR_OK;
 	if (buffer_size)
@@ -109,7 +109,7 @@ int ctr_fatfs_interface_write_sector(void *io, const void *buffer, size_t buffer
 	return result != FR_OK;
 }
 
-size_t ctr_fatfs_interface_disk_size(void *io)
+uint64_t ctr_fatfs_interface_disk_size(void *io)
 {
 	ctr_fatfs_interface *fatfs_io = io;
 	return f_size(fatfs_io->file);
