+++ src/ctr_io_interface.c
@@ -8,13 +8,13 @@
 
 #include <ctr9/io/ctr_io_interface.h>
 
-int ctr_io_read(void *io, void *buffer, size_t buffer_size, size_t position, size_t count)
+int ctr_io_read(void *io, void *buffer, size_t buffer_size, uint64_t position, size_t count)
 {
 	ctr_io_interface *base = io;
 	return base->read(io, buffer, buffer_size, position, count);
 }
 
-int ctr_io_write(void *io, const void *buffer, size_t buffer_size, size_t position)
+int ctr_io_write(void *io, const void *buffer, size_t buffer_size, uint64_t position)
 {
 	ctr_io_interface *base = io;
 	return base->write(io, buffer, buffer_size, position);
@@ -32,7 +32,7 @@ int ctr_io_write_sector(void *io, const void *buffer, size_t buffer_size, size_t
 	return base->write_sector(io, buffer, buffer_size, sector);
 }
 
-size_t ctr_io_disk_size(void *io)
+uint64_t ctr_io_disk_size(void *io)
 {
 	ctr_io_interface *base = io;
 	return base->disk_size(io);
